requirements.txt
pandas
plotly
pyyaml
streamlit
